Completed #9 Share Single Database Object across Classes
If we do not share the same database object and clicking on each Module AddMember and ViewMembers at the same time the program will crash because each one creates a new  DatabaseHandler object.
Make the constructor private and create a static method getInstance().
This is the same technique that we use for WebDriver instances creation.